feat: make chain source into mods
